requirements.txt
aiofiles==24.1.0
aiohttp==3.11.13
locustio==0.999
